+++ src/main/java/ru/itmo/wastemanagement/repository/StopEventRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.itmo.wastemanagement.entity.StopEvent;
 
 @Repository
-public interface StopEventRepository extends JpaRepository<StopEvent, Long> {
+public interface StopEventRepository extends JpaRepository<StopEvent, Integer> {
 }
